backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4
python-multipart==0.0.9
email-validator==2.2.0
google-auth==2.35.0
openai>=1.0,<2
python-dotenv==1.0.1
pandas==2.3.3
pdfplumber==0.11.9
python-docx==1.2.0
openpyxl==3.1.5
opencv-python==4.10.0.84
numpy==1.26.4
protobuf==4.25.3
mediapipe==0.10.21
ultralytics==8.3.23
